requirements.txt
djangorestframework>=3.14
djangorestframework-simplejwt
gunicorn>=20.1
psycopg2-binary>=2.9
celery>=5.3
redis>=4.5
django-cors-headers>=4.3
Pillow>=10.0
whitenoise
